requirements.txt
asgiref==3.8.1
Django==5.0.8
djangorestframework==3.15.2
dnspython==2.6.1
mysqlclient==2.2.4
pymongo==4.8.0
sqlparse==0.5.1
tzdata==2024.1
